fix : 시나리오 repository 데이터 타입 오류 수정

+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/scenario/feature/apilist/repository/ScenarioRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.sk.skala.axcalibur.scenario.feature.apilist.entity.ScenarioEntity;
 
 @Repository
-public interface ScenarioRepository extends JpaRepository<ScenarioEntity, Long> {
+public interface ScenarioRepository extends JpaRepository<ScenarioEntity, Integer> {
 
     /** 
      * 프로젝트 키로 시나리오 조회
